Refactoring crud system "

+++ src/Controller/PinController.php
@@ -2,7 +2,7 @@
 /*
  * @Author: your name
  * @Date: 2020-08-04 23:30:31
- * @LastEditTime: 2020-08-07 17:59:51
+ * @LastEditTime: 2020-08-07 21:00:04
  * @LastEditors: Please set LastEditors
  * @Description: In User Settings Edit
  * @FilePath: /cours-symfony-container/src/Controller/PinController.php
@@ -11,6 +11,7 @@
 namespace App\Controller;
 
 use App\Entity\Pin;
+use App\Form\PinType;
 use Faker\Provider\ar_JO\Text;
 use App\Repository\PinRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -56,23 +57,7 @@ class PinController extends AbstractController
     public function create(Request $request,EntityManagerInterface $em):Response
     {
         $pin = new Pin;
-        $form = $this->createFormBuilder($pin)
-            ->add('Title',TextType::class,[
-                'attr'=>[
-                    'required'=>true,
-                    'class'=>'form-control form-control-sm',
-                ],
-                'label'=>'Title'
-            ])
-            ->add('description',TextareaType::class,[
-                'attr'=>[
-                    'required'=>true,
-                    'class'=>'form-control form-control-sm'
-                ],
-                'label'=>'Description'
-            ])
-            ->getForm();
-
+        $form = $this->createForm(PinType::class, $pin);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             //$pin = $form->getData(); 
@@ -87,7 +72,7 @@ class PinController extends AbstractController
     }
 
     /**
-     * @Route("/pin/{id<[0-9]+>}/edit", name="app_edit_pin", methods={"GET","POST"})
+     * @Route("/pin/{id<[0-9]+>}/edit", name="app_edit_pin", methods={"GET","PUT"})
      *
      * @param Request $request
      * @param Pin $pin
@@ -96,46 +81,33 @@ class PinController extends AbstractController
     */
     public function edit(Request $request, Pin $pin, EntityManagerInterface $em ):Response
     {
-        $form = $this->createFormBuilder($pin)
-            ->add('Title',TextType::class,[
-                'attr'=>[
-                    'required'=>true,
-                    'class'=>'form-control form-control-sm',
-                ],
-                'label'=>'Title'
-            ])
-            ->add('description',TextareaType::class,[
-                'attr'=>[
-                    'required'=>true,
-                    'class'=>'form-control form-control-sm'
-                ],
-                'label'=>'Description'
-            ])
-            ->getForm();
-            
-            $form->handleRequest($request);
- 
-            if ($form->isSubmitted() && $form->isValid()) { 
-                $em->persist($pin);
-                $em->flush();
-                return $this->redirectToRoute('app_index_pin');
-             }
+        $form = $this->createForm(PinType::class, $pin,[
+            'method' =>'PUT'
+        ]);
+        $form->handleRequest($request);
+         if ($form->isSubmitted() && $form->isValid()) { 
+            $em->persist($pin);
+            $em->flush();
+            return $this->redirectToRoute('app_index_pin');
+         }
 
         return $this->render('pins/edit.html.twig',[
             'form'=>$form->createView()
         ]);
     }
 /**
- * @Route("pin/{id<[0-9]+>}/delete", name="app_delete_pin")
+ * @Route("pin/{id<[0-9]+>}/delete", name="app_delete_pin", methods={"GET","DELETE"})
  *
  * @param Pin $pin
  * @param EntityManagerInterface $em
  * @return void
  */
-    public function remove( Pin $pin, EntityManagerInterface $em)
+    public function remove(Request $request, Pin $pin, EntityManagerInterface $em)
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin.delete'.$pin->getId(),$request->request->get('csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
+        }
         return $this->redirectToRoute('app_index_pin');
     }
 }
